Add runtime flag to disable server-side rendering
