+++ .gitignore
@@ -67,3 +67,5 @@ env.json
 .elasticbeanstalk/*
 !.elasticbeanstalk/*.cfg.yml
 !.elasticbeanstalk/*.global.yml
+
+src/app/index.html
