+++ package.json
@@ -5,30 +5,41 @@
   "main": "index.js",
   "repository": "https://github.com/hollowverse/hollowverse.git",
   "scripts": {
-    "test": "run-p -l --aggregate-output validate-filenames lint-js lint-ts lint-css",
+    "test":
+      "run-p -l --aggregate-output validate-filenames lint-js lint-ts lint-css",
     "lint-css": "stylelint --syntax scss 'src/app/**/*.scss'",
     "lint-js": "eslint '**/*.js{,x}'",
-    "lint-ts/eslint": "eslint 'src/app/**/*.ts{,x}' --config src/app/.eslintrc.json --no-eslintrc",
+    "lint-ts/eslint":
+      "eslint 'src/app/**/*.ts{,x}' --config src/app/.eslintrc.json --no-eslintrc",
     "lint-ts/tslint": "tslint './*.ts' 'src/**/*.ts{,x}' --project .",
     "lint-ts": "run-p --aggregate-output 'lint-ts/*'",
-    "app/webpack/build": "REACT=1 NODE_ENV=production NO_PERF_CHECKS=1 node ./build.js",
+    "app/webpack/build":
+      "REACT=1 NODE_ENV=production NO_PERF_CHECKS=1 node ./build.js",
     "bundlesize": "bundlesize",
-    "app/build": "API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} run-s -l app/graphql/build app/webpack/build bundlesize",
+    "app/build":
+      "API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} run-s -l app/graphql/build app/webpack/build bundlesize",
     "app/analyze": "webpack-bundle-analyzer dist/stats.json dist/client",
-    "app/graphql/schema": "apollo-codegen introspect-schema ${API_ENDPOINT} --output ./graphqlSchema.json",
-    "app/graphql/types": "apollo-codegen generate 'src/app/**/*.{graphql,gql}' --target typescript --output src/app/api/types.ts",
-    "app/graphql/dev": "nodemon -q --watch src/app --ext graphql,gql --exec 'run-s app/graphql/schema app/graphql/types'",
+    "app/graphql/schema":
+      "apollo-codegen introspect-schema ${API_ENDPOINT} --output ./graphqlSchema.json",
+    "app/graphql/types":
+      "apollo-codegen generate 'src/app/**/*.{graphql,gql}' --target typescript --output src/app/api/types.ts",
+    "app/graphql/dev":
+      "nodemon -q --watch src/app --ext graphql,gql --exec 'run-s app/graphql/schema app/graphql/types'",
     "app/graphql/build": "run-s -l app/graphql/schema app/graphql/types",
-    "server/app-server/dev": "REACT=1 APP_SERVER_PORT=${APP_SERVER_PORT:-3001} HOT=1 NODE_ENV=development node --inspect=9231 -r 'source-map-support/register' -r 'ts-node/register' src/appServer.ts",
+    "server/app-server/dev":
+      "REACT=1 API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} APP_SERVER_PORT=${APP_SERVER_PORT:-3001} HOT=1 NODE_ENV=development node --inspect=9231 -r 'source-map-support/register' -r 'ts-node/register' src/appServer.dev.ts",
     "server/app-server/start": "node dist/appServer.js",
-    "server/main-server/dev": "PORT=${PORT:-8081} API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} nodemon -q --watch src --ignore src/app --ignore src/webpack --ext ts,tsx,json --exec 'node --inspect=0.0.0.0:${SERVER_DEBUG_PORT:-9232} -r 'ts-node/register' src/mainServer.ts'",
+    "server/main-server/dev":
+      "PORT=${PORT:-8081} API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} nodemon -q --watch src --ignore src/app --ignore src/webpack --ext ts,tsx,json --exec 'node --inspect=0.0.0.0:${SERVER_DEBUG_PORT:-9232} -r 'ts-node/register' src/mainServer.ts'",
     "remotedev": "remotedev --hostname=localhost --port=8000",
     "server/main-server/start": "node dist/mainServer.js",
     "server/build": "tsc --project ./src",
     "build": "npm-run-all -s clean -p app/build server/build",
     "start": "NODE_ENV=production run-p server/main-server/start",
-    "dev": "NODE_ENV=development API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} npm-run-all -l -s clean -p app/graphql/dev server/app-server/dev remotedev",
-    "clean": "del-cli './src/app/**/*.module.scss.d.ts' './dist' && mkdir -p ./dist",
+    "dev":
+      "NODE_ENV=development API_ENDPOINT=${API_ENDPOINT:-https://api.hollowverse.com/graphql} npm-run-all -l -s clean -p app/graphql/dev server/app-server/dev remotedev",
+    "clean":
+      "del-cli './src/app/**/*.module.scss.d.ts' './dist' && mkdir -p ./dist",
     "validate-filenames": "validate-filenames -c commonconfig.js",
     "lint-staged": "lint-staged"
   },
@@ -61,21 +72,10 @@
     }
   ],
   "lint-staged": {
-    "**/*.md": [
-      "prettier --write",
-      "git add"
-    ],
-    "**/*.{j,t}s{x,}": [
-      "prettier --write",
-      "git add"
-    ],
-    "**/*.json": [
-      "prettier --write",
-      "git add"
-    ],
-    "**/*.js{x,}": [
-      "eslint"
-    ],
+    "**/*.md": ["prettier --write", "git add"],
+    "**/*.{j,t}s{x,}": ["prettier --write", "git add"],
+    "**/*.json": ["prettier --write", "git add"],
+    "**/*.js{x,}": ["eslint"],
     "**/**/*.{s,}css": [
       "stylelint --syntax scss",
       "prettier --write",
@@ -134,7 +134,6 @@
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-flip-move": "^2.10.0",
-    "react-helmet": "^5.2.0",
     "react-helmet-async": "^0.0.4",
     "react-intersection-observer": "^3.0.2",
     "react-redux": "^5.0.6",
@@ -172,7 +171,6 @@
     "@types/domready": "^0.0.29",
     "@types/express": "^4.0.39",
     "@types/google.analytics": "^0.0.36",
-    "@types/helmet": "^0.0.37",
     "@types/http-proxy": "^1.12.0",
     "@types/lodash": "^4.14.68",
     "@types/loglevel": "^1.5.3",
@@ -209,6 +207,7 @@
     "babel-preset-minify": "^0.2.0",
     "bundlesize": "^0.15.3",
     "circular-dependency-plugin": "^3.0.0",
+    "css-chunks-html-webpack-plugin": "^0.1.0",
     "css-loader": "^0.28.4",
     "del-cli": "^1.1.0",
     "eslint": "^4.5.0",
@@ -225,6 +224,7 @@
     "file-loader": "^1.1.5",
     "graphql-loader": "^1.2.1",
     "html-loader": "^0.5.1",
+    "html-webpack-harddisk-plugin": "^0.1.0",
     "html-webpack-plugin": "^2.30.1",
     "http-proxy-middleware": "^0.17.4",
     "husky": "^0.15.0-beta.16",
