+++ src/app/index.client.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en" dir="ltr">
+  <head>
+    
+  </head>
+  <script>
+    window.__INITIAL_STATE__ = { };
+  </script>
+  <body>
+    <div style="min-height: 100%;" id="app"></div>
+  </body>
+</html>
