+++ src/app/index.html
@@ -2,8 +2,11 @@
 <html lang="en" dir="ltr">
   <head>
     
-  </head>
+  <link rel="apple-touch-icon" sizes="57x57" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-57x57.png"><link rel="apple-touch-icon" sizes="60x60" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-60x60.png"><link rel="apple-touch-icon" sizes="72x72" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-72x72.png"><link rel="apple-touch-icon" sizes="76x76" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-76x76.png"><link rel="apple-touch-icon" sizes="114x114" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-114x114.png"><link rel="apple-touch-icon" sizes="120x120" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-120x120.png"><link rel="apple-touch-icon" sizes="144x144" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-144x144.png"><link rel="apple-touch-icon" sizes="152x152" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-152x152.png"><link rel="apple-touch-icon" sizes="180x180" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-icon-180x180.png"><meta name="apple-mobile-web-app-capable" content="yes"><meta name="apple-mobile-web-app-status-bar-style" content="black-translucent"><meta name="apple-mobile-web-app-title" content="Hollowverse"><meta name="mobile-web-app-capable" content="yes"><meta name="theme-color" content="#fff"><meta name="application-name" content="Hollowverse"><link rel="icon" type="image/png" sizes="32x32" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/favicon-32x32.png"><link rel="icon" type="image/png" sizes="16x16" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/favicon-16x16.png"><link rel="shortcut icon" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/favicon.ico"><link rel="apple-touch-startup-image" media="(device-width: 320px) and (device-height: 480px) and (-webkit-device-pixel-ratio: 1)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-320x460.png"><link rel="apple-touch-startup-image" media="(device-width: 320px) and (device-height: 480px) and (-webkit-device-pixel-ratio: 2)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-640x920.png"><link rel="apple-touch-startup-image" media="(device-width: 320px) and (device-height: 568px) and (-webkit-device-pixel-ratio: 2)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-640x1096.png"><link rel="apple-touch-startup-image" media="(device-width: 375px) and (device-height: 667px) and (-webkit-device-pixel-ratio: 2)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-750x1294.png"><link rel="apple-touch-startup-image" media="(device-width: 414px) and (device-height: 736px) and (orientation: landscape) and (-webkit-device-pixel-ratio: 3)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-1182x2208.png"><link rel="apple-touch-startup-image" media="(device-width: 414px) and (device-height: 736px) and (orientation: portrait) and (-webkit-device-pixel-ratio: 3)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-1242x2148.png"><link rel="apple-touch-startup-image" media="(device-width: 768px) and (device-height: 1024px) and (orientation: landscape) and (-webkit-device-pixel-ratio: 1)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-748x1024.png"><link rel="apple-touch-startup-image" media="(device-width: 768px) and (device-height: 1024px) and (orientation: portrait) and (-webkit-device-pixel-ratio: 1)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-768x1004.png"><link rel="apple-touch-startup-image" media="(device-width: 768px) and (device-height: 1024px) and (orientation: landscape) and (-webkit-device-pixel-ratio: 2)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-1496x2048.png"><link rel="apple-touch-startup-image" media="(device-width: 768px) and (device-height: 1024px) and (orientation: portrait) and (-webkit-device-pixel-ratio: 2)" href="/static/icons-bd9db2e44d08ddc788936f771030d9a1/apple-touch-startup-image-1536x2008.png"><link href="/static/main.global.css" rel="stylesheet"><link href="/static/main.module.css" rel="stylesheet"><script type="text/javascript">window.__CSS_CHUNKS__ = {"NotablePerson":"/static/NotablePerson.module.css","SearchResults":"/static/SearchResults.module.css","Home":"/static/Home.module.css","PrivacyPolicy":"/static/PrivacyPolicy.module.css","About":"/static/About.module.css","main":"/static/main.global.css"}</script></head>
+  <script>
+    window.__INITIAL_STATE__ = { };
+  </script>
   <body>
     <div style="min-height: 100%;" id="app"></div>
-  </body>
+  <script type="text/javascript" src="/static/bootstrap.js"></script><script type="text/javascript" src="/static/vendor.js"></script><script type="text/javascript" src="/static/main.js"></script></body>
 </html>
