+++ src/webpack/webpack.config.client.js
@@ -4,6 +4,8 @@ const webpackMerge = require('webpack-merge');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 const FaviconsWebpackPlugin = require('favicons-webpack-plugin');
 const ExtractCssChunks = require('extract-css-chunks-webpack-plugin');
+const HtmlWebpackHarddiskPlugin = require('html-webpack-harddisk-plugin');
+const CssChunkHashPlugin = require('css-chunks-html-webpack-plugin');
 
 const NameAllModulesPlugin = require('name-all-modules-plugin');
 
@@ -110,10 +112,13 @@ const clientSpecificConfig = {
       inject: true,
     }),
 
+    new CssChunkHashPlugin(),
+
     new HtmlWebpackPlugin({
-      template: path.join(srcDirectory, 'index.html'),
+      template: path.join(srcDirectory, 'index.client.html'),
       filename: 'index.html',
       inject: 'body',
+      alwaysWriteToDisk: true,
       minify: isProd
         ? {
             html5: true,
@@ -122,6 +127,10 @@ const clientSpecificConfig = {
             collapseWhitespace: true,
           }
         : false,
+    }),
+
+    new HtmlWebpackHarddiskPlugin({
+      outputPath: ifDev(path.resolve(__dirname, '../app')),
     }),
 
     // Required for debugging in development and for long-term caching in production
